Translating Submission Statuses and refactor
